Uses scalamock instead of mockito and moves some tests to shared proj
