+++ github4s/shared/src/test/scala/github4s/unit/GHStatusesSpec.scala
@@ -20,64 +20,52 @@ import cats.free.Free
 import github4s.GHStatuses
 import github4s.GithubResponses.{GHResponse, GHResult}
 import github4s.app.GitHub4s
-import github4s.free.algebra.StatusOps
 import github4s.free.domain.{CombinedStatus, Status}
-import github4s.utils.TestUtils
-import org.mockito.ArgumentMatchers.any
-import org.mockito.Mockito._
-import org.scalatest.{FlatSpec, Matchers}
-import org.scalatest.mockito.MockitoSugar.mock
+import github4s.utils.BaseSpec
 
-class GHStatusesSpec extends FlatSpec with Matchers with TestUtils {
+class GHStatusesSpec extends BaseSpec {
+
   "GHStatuses.get" should "call to StatusesOps with the right parameters" in {
     val response: Free[GitHub4s, GHResponse[CombinedStatus]] =
       Free.pure(Right(GHResult(combinedStatus, okStatusCode, Map.empty)))
 
-    val statusesOps = mock[StatusOps[GitHub4s]]
-    when(statusesOps.getCombinedStatus(any[String], any[String], any[String], any[Option[String]]))
-      .thenReturn(response)
-
-    val token      = Some("token")
-    val ghStatuses = new GHStatuses(token)(statusesOps)
+    val statusesOps = mock[StatusOpsTest]
+    (statusesOps.getCombinedStatus _)
+      .expects(validRepoOwner, validRepoName, validRefSingle, sampleToken)
+      .returns(response)
+    val ghStatuses = new GHStatuses(sampleToken)(statusesOps)
     ghStatuses.getCombinedStatus(validRepoOwner, validRepoName, validRefSingle)
-
-    verify(statusesOps).getCombinedStatus(validRepoOwner, validRepoName, validRefSingle, token)
   }
 
   "GHStatuses.list" should "call to StatusesOps with the right parameters" in {
     val response: Free[GitHub4s, GHResponse[List[Status]]] =
       Free.pure(Right(GHResult(List(status), okStatusCode, Map.empty)))
 
-    val statusesOps = mock[StatusOps[GitHub4s]]
-    when(statusesOps.listStatuses(any[String], any[String], any[String], any[Option[String]]))
-      .thenReturn(response)
-
-    val token      = Some("token")
-    val ghStatuses = new GHStatuses(token)(statusesOps)
+    val statusesOps = mock[StatusOpsTest]
+    (statusesOps.listStatuses _)
+      .expects(validRepoOwner, validRepoName, validRefSingle, sampleToken)
+      .returns(response)
+    val ghStatuses = new GHStatuses(sampleToken)(statusesOps)
     ghStatuses.listStatuses(validRepoOwner, validRepoName, validRefSingle)
-
-    verify(statusesOps).listStatuses(validRepoOwner, validRepoName, validRefSingle, token)
   }
 
   "GHStatuses.create" should "call to StatusesOps with the right parameters" in {
     val response: Free[GitHub4s, GHResponse[Status]] =
       Free.pure(Right(GHResult(status, createdStatusCode, Map.empty)))
 
-    val statusesOps = mock[StatusOps[GitHub4s]]
-    when(
-      statusesOps.createStatus(
-        any[String],
-        any[String],
-        any[String],
-        any[String],
-        any[Option[String]],
-        any[Option[String]],
-        any[Option[String]],
-        any[Option[String]]))
-      .thenReturn(response)
-
-    val token      = Some("token")
-    val ghStatuses = new GHStatuses(token)(statusesOps)
+    val statusesOps = mock[StatusOpsTest]
+    (statusesOps.createStatus _)
+      .expects(
+        validRepoOwner,
+        validRepoName,
+        validCommitSha,
+        validStatusState,
+        None,
+        None,
+        None,
+        sampleToken)
+      .returns(response)
+    val ghStatuses = new GHStatuses(sampleToken)(statusesOps)
     ghStatuses
       .createStatus(
         validRepoOwner,
@@ -87,15 +75,5 @@ class GHStatusesSpec extends FlatSpec with Matchers with TestUtils {
         None,
         None,
         None)
-
-    verify(statusesOps).createStatus(
-      validRepoOwner,
-      validRepoName,
-      validCommitSha,
-      validStatusState,
-      None,
-      None,
-      None,
-      token)
   }
 }
